+++ data-raw/build_test_fixtures.R
@@ -6,7 +6,7 @@ setwd(here::here())
 
 mfl_conn <- ffscrapr::mfl_connect(season = 2021, league_id = 22627)
 
-mfl_scoring_history <- ffscrapr::ff_scoringhistory(mfl_conn, 2015:2020) %>%
+mfl_scoring_history <- ffscrapr::ff_scoringhistory(mfl_conn, 2019:2020) %>%
   select(season, week, gsis_id, sportradar_id, mfl_id, player_name, pos, team, points)
 
 mfl_franchises <- ffs_franchises(mfl_conn)
@@ -24,7 +24,7 @@ saveRDS(mfl_lineup_constraints, "inst/cache/mfl_lineup_constraints.rds")
 
 sleeper_conn <- ff_connect(platform = "sleeper", league_id = "652718526494253056", season = 2021)
 
-sleeper_scoring_history <- ffscrapr::ff_scoringhistory(sleeper_conn, 2015:2020) %>%
+sleeper_scoring_history <- ffscrapr::ff_scoringhistory(sleeper_conn, 2019:2020) %>%
   select(season, week, gsis_id, sportradar_id, sleeper_id, player_name, pos, team, points)
 
 sleeper_franchises <- ffs_franchises(sleeper_conn)
@@ -40,7 +40,7 @@ saveRDS(sleeper_lineup_constraints, "inst/cache/sleeper_lineup_constraints.rds")
 
 fleaflicker_conn <- fleaflicker_connect(2020, 206154)
 
-fleaflicker_scoring_history <- ffscrapr::ff_scoringhistory(fleaflicker_conn, 2015:2020) %>%
+fleaflicker_scoring_history <- ffscrapr::ff_scoringhistory(fleaflicker_conn, 2019:2020) %>%
   select(season, week, gsis_id, sportradar_id, player_name, pos, team, points)
 
 fleaflicker_franchises <- ffs_franchises(fleaflicker_conn)
@@ -56,7 +56,7 @@ saveRDS(fleaflicker_lineup_constraints, "inst/cache/fleaflicker_lineup_constrain
 
 espn_conn <- espn_connect(season = 2020, league_id = 899513)
 
-espn_scoring_history <- ffscrapr::ff_scoringhistory(espn_conn, 2015:2020) %>%
+espn_scoring_history <- ffscrapr::ff_scoringhistory(espn_conn, 2019:2020) %>%
   select(season, week, gsis_id, sportradar_id, player_name, pos, team, points)
 
 espn_franchises <- ffs_franchises(espn_conn)
@@ -70,8 +70,10 @@ saveRDS(espn_franchises, "inst/cache/espn_franchises.rds")
 saveRDS(espn_rosters, "inst/cache/espn_rosters.rds")
 saveRDS(espn_lineup_constraints, "inst/cache/espn_lineup_constraints.rds")
 
-latest_rankings <- ffs_latest_rankings()
+latest_rankings <- ffs_latest_rankings(type = "draft")
+latest_rankings_week <- ffs_latest_rankings(type = "week")
 saveRDS(latest_rankings, "inst/cache/latest_rankings.rds")
+saveRDS(latest_rankings_week, "inst/cache/latest_rankings_week.rds")
 
 adp_outcomes <- ffs_adp_outcomes(
   scoring_history = mfl_scoring_history,
@@ -79,14 +81,28 @@ adp_outcomes <- ffs_adp_outcomes(
 )
 saveRDS(adp_outcomes, "inst/cache/adp_outcomes.rds")
 
+adp_outcomes_week <- ffs_adp_outcomes_week(
+  scoring_history = mfl_scoring_history,
+  pos_filter = c("QB","RB","WR")
+)
+saveRDS(adp_outcomes_week, "inst/cache/adp_outcomes_week.rds")
+
 projected_scores <- ffs_generate_projections(
   adp_outcomes = adp_outcomes,
   latest_rankings = latest_rankings,
   n_seasons = 2,
-  weeks = 1:5,
+  weeks = 1:3,
   rosters = mfl_rosters
 )
 saveRDS(projected_scores, "inst/cache/projected_scores.rds")
+
+projected_scores_week <- ffs_generate_projections_week(
+  adp_outcomes = adp_outcomes_week,
+  latest_rankings = latest_rankings_week,
+  n = 5,
+  rosters = mfl_rosters
+)
+saveRDS(projected_scores_week, "inst/cache/projected_scores_week.rds")
 
 roster_scores <- ffs_score_rosters(
   projected_scores = projected_scores,
@@ -104,13 +120,16 @@ saveRDS(optimal_scores, "inst/cache/optimal_scores.rds")
 schedules <- ffs_build_schedules(
   n_teams = NULL,
   n_seasons = 2,
-  n_weeks = 5,
+  n_weeks = 3,
   franchises = mfl_franchises
 )
 
 saveRDS(schedules, "inst/cache/schedules.rds")
 
 foureight <- mfl_connect(2021, 22627)
-foureight_sim <- ff_simulate(foureight, n_seasons = 25)
+foureight_sim <- ff_simulate(foureight, n_seasons = 10)
 
 saveRDS(foureight_sim, "inst/cache/foureight_sim.rds")
+
+foureight_sim_week <- ff_simulate_week(foureight,n = 10)
+saveRDS(foureight_sim_week, "inst/cache/foureight_sim_week.rds")